Add common implementation of make_valid_dids to utils.
